strategic_mode_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
httpx==0.28.1
redis==5.2.1
celery==5.4.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
alembic==1.14.1
prometheus-client==0.21.1
structlog==24.5.0